+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.1-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.8.0...@quiet/mobile@1.8.1-alpha.0) (2023-08-01)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+# [1.8.0](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@1.6.1-alpha.0...@quiet/mobile@1.8.0) (2023-07-28)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+# [1.7.0](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@1.6.1-alpha.0...@quiet/mobile@1.7.0) (2023-07-28)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.6.1-alpha.0](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@1.6.0-alpha.0...@quiet/mobile@1.6.1-alpha.0) (2023-07-28)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+# [1.6.0-alpha.0](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@1.5.0-alpha.0...@quiet/mobile@1.6.0-alpha.0) (2023-07-26)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
 # [1.5.0-alpha.0](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@1.4.2-alpha.1...@quiet/mobile@1.5.0-alpha.0) (2023-07-18)
 
 **Note:** Version bump only for package @quiet/mobile